Enable Dynamic Colors on Android 12+ and update status bar color
This commit enables the use of dynamic colors on Android 12 and newer devices. It also updates the status bar color to be transparent and adjusts the status bar's appearance to be light or dark based on the current theme. For learning purposes, dynamic colors are set as off by default, but they can be easily toggled on by setting `dynamicColor` to `true` in `AppTheme`.